+++ test-support/helpers/ember-tb-test-helpers/test-helpers.js
@@ -1,0 +1,1 @@
+import 'ember-tb-test-helpers/within';