+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 master
 ------
 
+* Introduce `within` helper, for scoping helpers to a selector
+
 0.0.1
 -----